refactor: split private paykit repo
